pyproject.toml
[tool.poetry]
name = "code_gen"
version = "0.1.0"
description = "A tool for generating blog code samples"
authors = ["Kenny Sato <kennysat@amazon.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# aiohttp = "^3.9.5"
aiohttp = "^3.10.2"
aiosignal = "^1.3.1"
annotated-types = "^0.7.0"
anthropic = "^0.28.0"
anyio = "^4.4.0"
appnope = "^0.1.4"
asttokens = "^2.4.1"
attrs = "^23.2.0"
beautifulsoup4 = "^4.12.3"
boto3 = "^1.34.122"
botocore = "^1.34.122"
bs4 = "^0.0.2"
certifi = "^2024.6.2"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
comm = "^0.2.2"
contourpy = "^1.2.1"
cycler = "^0.12.1"
dataclasses-json = "^0.6.6"
debugpy = "^1.8.1"
decorator = "^5.1.1"
defusedxml = "^0.7.1"
distro = "^1.9.0"
docstring_parser = "^0.16"
executing = "^2.0.1"
filelock = "^3.14.0"
fonttools = "^4.53.0"
frozenlist = "^1.4.1"
fsspec = "^2024.6.0"
graphviz = "^0.20.3"
greenlet = "^3.0.3"
h11 = "^0.14.0"
httpcore = "^1.0.5"
httpx = "^0.27.0"
huggingface-hub = "^0.23.2"
idna = "^3.7"
instructor = "^1.3.2"
ipykernel = "^6.29.4"
ipython = "^8.25.0"
jedi = "^0.19.1"
jiter = "^0.4.1"
jmespath = "^1.0.1"
joblib = "^1.4.2"
jsonpatch = "^1.33"
jsonpointer = "^2.4"
jupyter_client = "^8.6.2"
jupyter_core = "^5.7.2"
jupyter_server = "^2.14.1"
kiwisolver = "^1.4.5"
# langchain = "^0.2.3"
langchain = "^0.2.9"
langchain-anthropic = "^0.1.15"
langchain-aws = "^0.1.6"
langchain-community = "^0.2.4"
langchain-core = "^0.2.4"
langchain-openai = "^0.1.8"
langchain-text-splitters = "^0.2.1"
langgraph = "^0.0.65"
langsmith = "^0.1.69"
markdown-it-py = "^3.0.0"
marshmallow = "^3.21.2"
matplotlib = "^3.9.0"
matplotlib-inline = "^0.1.7"
mdurl = "^0.1.2"
multidict = "^6.0.5"
mypy-extensions = "^1.0.0"
nest-asyncio = "^1.6.0"
numpy = "^1.26.4"
openai = "^1.31.0"
orjson = "^3.10.3"
packaging = "^23.2"
parso = "^0.8.4"
pexpect = "^4.9.0"
pillow = "^10.3.0"
platformdirs = "^4.2.2"
prompt_toolkit = "^3.0.46"
psutil = "^5.9.8"
ptyprocess = "^0.7.0"
pure-eval = "^0.2.2"
pydantic = "^2.7.3"
pydantic_core = "^2.18.4"
Pygments = "^2.18.0"
pyparsing = "^3.1.2"
python-dateutil = "^2.9.0.post0"
PyYAML = "^6.0.1"
pyzmq = "^26.0.3"
regex = "^2024.5.15"
requests = "^2.32.3"
rich = "^13.7.1"
s3transfer = "^0.10.1"
scikit-learn = "^1.5.0"
scipy = "^1.13.1"
shellingham = "^1.5.4"
six = "^1.16.0"
sniffio = "^1.3.1"
soupsieve = "^2.5"
SQLAlchemy = "^2.0.30"
stack-data = "^0.6.3"
tenacity = "^8.3.0"
threadpoolctl = "^3.5.0"
tiktoken = "^0.7.0"
tokenizers = "^0.19.1"
tornado = "^6.4"
tqdm = "^4.66.4"
traitlets = "^5.14.3"
typer = "^0.12.3"
typing-inspect = "^0.9.0"
typing_extensions = "^4.12.1"
urllib3 = "^2.2.1"
wcwidth = "^0.2.13"
yarl = "^1.9.4"
pandas = "^2.2.2"
jupyter-server = "^2.14.1"
xmltodict = "^0.12.0"
pydantic-settings = "^2.3.4"
bandit = "^1.7.10"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
black = "^22.12.0"
flake8 = "^6.0.0"
mypy = "^0.991"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[virtualenvs]
create = true
in-project = true

[tool.bandit]
exclude_dirs = [".venv", "breast_cancer/breast_cancer_venv"]